requirements.txt
Pillow==7.0.0
pynput==1.6.6
PyQt5==5.9.2
sip==4.19.8
six==1.14.0
